+++ install.sh
@@ -5,6 +5,7 @@
 # Department of Bioinformatics and Biostatistics, Shanghai Jiao Tong University
 # Contact information: buikien.dp@sjtu.edu.cn, ccwei@sjtu.edu.cn
 #
+set -e
 
 FSCRPT=$(readlink -f "$0")
 LDIR=$(dirname "$FSCRPT")
@@ -14,15 +15,17 @@ NB=`wc -l < $LDIR/.tmp`
 
 
 
-g++ -std=c++11 -o $LDIR/src/genSpecies $LDIR/src/genSpecies.cpp -O3
+g++ -std=c++11 -o $LDIR/src/genPathFiles $LDIR/src/genPathFiles.cpp -O3
 g++ -std=c++11 -o $LDIR/src/buildDB  $LDIR/src/DTB_compress.cpp -O3
 g++ -std=c++11 -o $LDIR/src/uniqueDB  $LDIR/src/DTB_unique.cpp -O3
 g++ -std=c++11 -o $LDIR/src/translate  $LDIR/src/translate.cpp -O3
 g++ -fopenmp -std=c++11 -o $LDIR/src/classify  $LDIR/src/classify.cpp $LDIR/src/seqreader.cpp -O3
 g++ -fopenmp -std=c++11 -o $LDIR/src/classifyEM  $LDIR/src/classifyEM.cpp $LDIR/src/seqreader.cpp -O3
 
+mv $LDIR/src/genPathFiles $LDIR/
 mv $LDIR/src/buildDB $LDIR/
 mv $LDIR/src/uniqueDB $LDIR/
 mv $LDIR/src/translate $LDIR/
 mv $LDIR/src/classify $LDIR/
 mv $LDIR/src/classifyEM $LDIR/
+
