+++ src/DTB_unique.cpp
@@ -8,82 +8,10 @@
 */
 
 #include <bits/stdc++.h>
-#define LL long long
-#define ULL unsigned long long
-#define FOR(i,a,b) for(uint32_t i=a;i<=b;i++)
-#define FO(i,a,b) for(uint32_t i=a;i<b;i++)
-#define DEBUG(a) {cerr << #a << ": " << (a) << endl; fflush(stderr); }
-
-using namespace std;
-template<class T> int getbit(T s, int i) { return (s >> i) & 1; }
-template<class T> T onbit(T s, int i) { return s | (T(1) << i); }
-template<class T> T offbit(T s, int i) { return s & (~(T(1) << i)); }
-template<class T> int cntbit(T s) { return __builtin_popcount(s);}
-typedef pair<uint32_t, uint32_t> II32;
-
-
-/***************************************************************/
-#include <sys/time.h>
-#include <sys/resource.h>
-inline void printRam() {
-    struct rusage ru;
-    getrusage(RUSAGE_SELF, &ru);
-    cerr << "Max ram (in kilobytes): " << ru.ru_maxrss << endl;
-}
-
-namespace Time{
-    double start_time, time_limit;
-    static double last_call = 0;
-    int get_time_calls = 0;
-
-    double get_time() {
-        get_time_calls++;
-        timeval tv;
-        gettimeofday(&tv, 0);
-        return tv.tv_sec+tv.tv_usec*1e-6;
-    }
-
-    void print_time(string s) {
-#ifdef LOCAL
-    double x = get_time();
-    fprintf(stderr,"%s cur=%.6lf lap=%.6lf\n",s.c_str(),x,x-last_call);
-    last_call = x;
-#endif
-    }
-
-    void init_time() {
-        start_time = get_time();
-        last_call = start_time;
-    }
-}
-/******************************************************************/
-
-
-
-
-inline int get_code(char c) {
-    if (c == 'A') return 0;
-    if (c == 'C') return 1;
-    if (c == 'G') return 2;
-    if (c == 'T') return 3;
-    return 0;
-}
-
-inline char print_code(int c) {
-    if (c == 0) return 'A';
-    if (c == 1) return 'C';
-    if (c == 2) return 'G';
-    if (c == 3) return 'T';
-    return 'N';
-}
-
-
-
-
+#include "helpers.h"
 
 const int KMER = 32, PREFIX = 14, MAXBIT = 1<<(2*PREFIX);
 uint64_t RIGHT16 = 4294967295;
-
 
 class HashTable {
 public:
@@ -104,14 +32,14 @@ public:
 
     void sortData() {
         int cntRemove = 0;
-        FO (i,0,MAXBIT) {
+        for (size_t i = 0; i < MAXBIT; ++i) {
             if (mTable[i].size() > 2) {
                 sort(mTable[i].begin(), mTable[i].end());
                 Vtmp.clear();
                 Vtmp.shrink_to_fit();
                 int freq = 1;
                 mTable[i].push_back(II32(-1,-1));
-                FO (j,0,mTable[i].size()-1) {
+                for (size_t j = 0; j < mTable[i].size()-1; ++j) {
                     if (mTable[i][j].first != mTable[i][j+1].first) {
                         if(freq == 1)
                             Vtmp.push_back(mTable[i][j]);
@@ -132,23 +60,9 @@ public:
         }
         DEBUG(cntRemove);
     }
-
 };
 HashTable HT;
 
-
-uint64_t reverseMask3(uint64_t _ikmer, int m_k) {
-    uint64_t _ikmerR = _ikmer;
-    // The following 6 lines come from Jellyfish source code
-    _ikmerR = ((_ikmerR >> 2)  & 0x3333333333333333UL) | ((_ikmerR & 0x3333333333333333UL) << 2);
-    _ikmerR = ((_ikmerR >> 4)  & 0x0F0F0F0F0F0F0F0FUL) | ((_ikmerR & 0x0F0F0F0F0F0F0F0FUL) << 4);
-    _ikmerR = ((_ikmerR >> 8)  & 0x00FF00FF00FF00FFUL) | ((_ikmerR & 0x00FF00FF00FF00FFUL) << 8);
-    _ikmerR = ((_ikmerR >> 16) & 0x0000FFFF0000FFFFUL) | ((_ikmerR & 0x0000FFFF0000FFFFUL) << 16);
-    _ikmerR = ( _ikmerR >> 32                        ) | (_ikmerR                        << 32);
-    _ikmerR = (((uint64_t)-1) - _ikmerR) >> (64 - (m_k << 1));
-    if(_ikmer < _ikmerR) return _ikmer;
-    return _ikmerR;
-}
 
 void usage() {
     cerr << "./DTB_unique database.txt FinalDTB\n";
@@ -194,13 +108,13 @@ int main(int argc, char **argv) {
 
     DEBUG("Writing the final database");
     uint64_t cntNum = 0, HTsize = 0;
-    FO (i,0,MAXBIT) {
+    for (size_t i = 0; i < MAXBIT; ++i) {
         HTsize += HT.mTable[i].size();
     }
     DEBUG(HTsize);
     foutSize.write((char *) &HTsize, sizeof(HTsize));
 
-	FO (i,0,MAXBIT) {
+	for (size_t i = 0; i < MAXBIT; ++i) {
 	    uint32_t sz = HT.mTable[i].size();
 	    foutSize.write((char *) &sz, sizeof(sz));
 	    if (sz == 0)
